Moved some stuff to utils
