+++ src/index.ts
@@ -1,5 +1,5 @@
 import { app, port } from './app.js'
-import { databaseInit } from './databaseConnection'
+import { databaseInit } from './utils/databaseConnection.js'
 
 databaseInit()
   .then(() => {
