+++ src/routes/joinGame.ts
@@ -1,9 +1,9 @@
-import { getClient } from '../databaseConnection'
+import { getClient } from '../utils/databaseConnection'
 import { celebrate, Joi, Segments } from 'celebrate'
 import sha256 from 'crypto-js/sha256'
 import { getMessaging } from 'firebase-admin/messaging'
 import { type gameSettings } from '../app'
-import { verifyFCMToken } from '../firebase'
+import { verifyFCMToken } from '../utils/firebase'
 import express, { type Router } from 'express'
 
 const router: Router = express.Router()
