+++ src/tests/databaseConnection.test.ts
@@ -1,4 +1,4 @@
-import { databaseInit } from '../databaseConnection'
+import { databaseInit } from '../utils/databaseConnection'
 
 test('Database connection', async () => {
   await expect(databaseInit()).resolves.not.toThrow()
