+++ src/utils/databaseConnection.ts
@@ -1,5 +1,5 @@
 import pg from 'pg'
-import { user, password } from './secrets'
+import { user, password } from '../secrets'
 // pg is a CommonJS module, so we have to do it this way for the import to work
 export const { Client } = pg
